+++ modules/textures/src/index.d.ts
@@ -1,3 +1,8 @@
+export {BasisLoader, BasisWorkerLoader} from './basis-loader';
+export {CompressedTextureLoader} from './compressed-texture-loader';
+export {CrunchLoader, CrunchWorkerLoader} from './crunch-loader';
+
+export {CompressedTextureWriter} from './compressed-texture-writer';
 
 // Texture Loading API
 export {loadImageTexture} from './lib/texture-api/load-image';